+ Training on windows

+++ examples/regressor.c
@@ -1,5 +1,6 @@
 #include "darknet.h"
-#include <sys/time.h>
+#include "utils.h"
+#include <time.h>
 #include <assert.h>
 
 void train_regressor(char *datacfg, char *cfgfile, char *weightfile, int *gpus, int ngpus, int clear)
@@ -64,7 +65,7 @@ void train_regressor(char *datacfg, char *cfgfile, char *weightfile, int *gpus, 
     args.d = &buffer;
     load_thread = load_data(args);
 
-    int epoch = (*net->seen)/N;
+    int epoch = (int)((*net->seen)/N);
     while(get_current_batch(net) < net->max_batches || net->max_batches == 0){
         time=clock();
 
@@ -87,7 +88,7 @@ void train_regressor(char *datacfg, char *cfgfile, char *weightfile, int *gpus, 
 #endif
         if(avg_loss == -1) avg_loss = loss;
         avg_loss = avg_loss*.9 + loss*.1;
-        printf("%ld, %.3f: %f, %f avg, %f rate, %lf seconds, %ld images\n", get_current_batch(net), (float)(*net->seen)/N, loss, avg_loss, get_current_rate(net), sec(clock()-time), *net->seen);
+        printf("%ld, %.3f: %f, %f avg, %f rate, %lf seconds, %ld images\n", (long)get_current_batch(net), (float)(*net->seen)/N, loss, avg_loss, get_current_rate(net), sec(clock()-time), (long)*net->seen);
         free_data(train);
         if(*net->seen/N > epoch){
             epoch = *net->seen/N;
@@ -209,7 +210,7 @@ void run_regressor(int argc, char **argv)
     int ngpus = 0;
     if(gpu_list){
         printf("%s\n", gpu_list);
-        int len = strlen(gpu_list);
+        int len = (int)strlen(gpu_list);
         ngpus = 1;
         int i;
         for(i = 0; i < len; ++i){
